requirements.txt
python-twitter==2.0
pytz==2013.7
soundcloud==0.4.1
wordnik==2.1.2